third_party/WebKit/LayoutTests/http/tests/inspector/compiler-source-mapping-debug.html

<html>
<head>
<script src="inspector-test.js"></script>
<script src="debugger-test.js"></script>
<script src="console-test.js"></script>
<script src="network-test.js"></script>
<script src="resources/compiled.js"></script>
<script src="resources/compiled-with-wrong-source-map-url.js"></script>

<script>

function clickButton()
{
    document.getElementById('test').click();
}

function test()
{
    InspectorTest.runDebuggerTestSuite([
        function testSetBreakpoint(next)
        {
            WebInspector.settings.jsSourceMapsEnabled.get = function() { return true; };
            InspectorTest.reloadPage(pageReloaded);

            function pageReloaded()
            {
                InspectorTest.showScriptSource("source1.js", didShowSource);
            }

            function didShowSource(sourceFrame)
            {
                InspectorTest.addResult("Script source was shown.");
                InspectorTest.setBreakpoint(sourceFrame, 14, "", true);
                InspectorTest.waitUntilPaused(paused);
                InspectorTest.evaluateInPage("setTimeout(clickButton, 0)");
            }

            function paused(callFrames)
            {
                InspectorTest.captureStackTrace(callFrames);
                next();
            }
        },

        function testNo404InConsole(next)
        {
            function callbackWhenSourceMapRequestIsAvailable(callback)
            {
                var requests = InspectorTest.networkRequests();
                for (var i = 0; i < requests.length; ++i) {
                    if (requests[i].url.indexOf("source-map.json_not_found") === -1 || !requests[i].finished)
                        continue;
                    var request = requests[i];
                    callback(request);
                    return;
                }
                InspectorTest.addSniffer(WebInspector.NetworkManager.prototype, "_finishNetworkRequest", callbackWhenSourceMapRequestIsAvailable);
            }

            callbackWhenSourceMapRequestIsAvailable(step2);

            function step2(request)
            {
                InspectorTest.assertEquals(request.statusCode, 404);
                InspectorTest.dumpConsoleMessages();
                next();
            }
        }
    ]);
};

</script>

</head>

<body onload="runTest()">
<p>Tests installing compiler source map in scripts panel.</p>
</body>
</html>
