third_party/WebKit/LayoutTests/http/tests/inspector/compiler-source-mapping-debug-expected.txt

CONSOLE MESSAGE: line 2: button clicked!
Tests installing compiler source map in scripts panel.



Running: testSetBreakpoint
Page reloaded.
Script source was shown.
Script execution paused.
Call stack:
    0) handleClick (compiled.js:2)
    1) clickButton (compiler-source-mapping-debug.html:14)

Running: testNo404InConsole
Script execution resumed.
